refactor: modularize rest bindings
