+++ pkgs/standards/autoapi/autoapi/v3/bindings/rest/__init__.py
@@ -1,0 +1,3 @@
+from .attach import build_router_and_attach
+
+__all__ = ["build_router_and_attach"]
