+++ pkgs/standards/autoapi/autoapi/v3/bindings/rest/attach.py
@@ -1,0 +1,26 @@
+from __future__ import annotations
+
+from types import SimpleNamespace
+from typing import Optional, Sequence
+
+from .common import OpSpec, _Key, logger
+from .router import _build_router
+
+
+def build_router_and_attach(
+    model: type, specs: Sequence[OpSpec], *, only_keys: Optional[Sequence[_Key]] = None
+) -> None:
+    """
+    Build a Router for the model and attach it to `model.rest.router`.
+    For simplicity and correctness with FastAPI, we **rebuild the entire router**
+    on each call (FastAPI does not support removing individual routes cleanly).
+    """
+    router = _build_router(model, specs)
+    rest_ns = getattr(model, "rest", None) or SimpleNamespace()
+    rest_ns.router = router
+    setattr(model, "rest", rest_ns)
+    logger.debug(
+        "rest: %s router attached with %d routes",
+        model.__name__,
+        len(getattr(router, "routes", []) or []),
+    )
