+++ pkgs/standards/autoapi/tests/unit/test_bulk_body_annotation.py
@@ -1,6 +1,6 @@
 from typing import get_args, get_origin
 
-from autoapi.v3.bindings import rest as rest_binding
+from autoapi.v3.bindings.rest.collection import _make_collection_endpoint
 from autoapi.v3.orm.mixins import BulkCapable, GUIDPk
 from autoapi.v3.ops import OpSpec
 from autoapi.v3.orm.tables import Base
@@ -15,7 +15,7 @@ def test_bulk_create_body_annotation_is_list() -> None:
         name = Column(String, nullable=False)
 
     sp = OpSpec(alias="bulk_create", target="bulk_create")
-    endpoint = rest_binding._make_collection_endpoint(
+    endpoint = _make_collection_endpoint(
         Widget, sp, resource="widget", db_dep=lambda: None
     )
     body_ann = endpoint.__annotations__["body"]
