+++ pkgs/standards/autoapi/tests/unit/test_db_dependency.py
@@ -1,4 +1,4 @@
-from autoapi.v3.bindings.rest import _build_router
+from autoapi.v3.bindings.rest.router import _build_router
 from autoapi.v3.ops import OpSpec
 from autoapi.v3.orm.tables import Base
 from autoapi.v3.orm.mixins import GUIDPk
