+++ pkgs/standards/autoapi/tests/unit/test_op_ctx_arity_paths.py
@@ -1,7 +1,7 @@
 from autoapi.v3.types import App
 from autoapi.v3 import op_ctx
 from autoapi.v3.decorators import collect_decorated_ops
-from autoapi.v3.bindings.rest import _build_router
+from autoapi.v3.bindings.rest.router import _build_router
 from autoapi.v3.orm.tables import Base
 from autoapi.v3.orm.mixins import GUIDPk
 from autoapi.v3.types import Column, String
