+++ pkgs/standards/autoapi/tests/unit/test_security_per_route.py
@@ -2,7 +2,7 @@ from autoapi.v3.ops import OpSpec
 from autoapi.v3.autoapp import AutoApp
 from autoapi.v3.orm.tables import Base
 from autoapi.v3.orm.mixins import GUIDPk
-from autoapi.v3.bindings.rest import _build_router
+from autoapi.v3.bindings.rest.router import _build_router
 from fastapi import FastAPI, Security
 from fastapi.security import HTTPBearer
